Add files via upload
It is maybe nitpicking, but  AgileArtificialIntelligence.github.io/book/06-Data/figures/networkTopology.png  shows an inconsistency with the method NNetwork>>viewNetworkIn:  earlier in the text. The tab in the png is called "network" but the code calls it "Network".
This is the corrected picture. Hope it works for the book.