Fixing small mistakes in 02-chapter2
Hello,
I found a little grammatical error in this chapter: "the teacher can gives" -> "the teacher can give"
Also the code of the Figures 11, 12 of the section 1.9 does not match with the code examples, the text (that I changed) and Figure 13, although they do the same. I changed the text under the Figure 11 to patch those differences, but I would suggest a change for Figures 11 and 12, or a change for all code examples and even Figure 13.